% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatezones.R
\docType{data}
\name{climatezones}
\alias{climatezones}
\title{Koppen-Geiger Climate Zones reference table.}
\format{A data frame with 92416 rows and 3 variables:
\describe{
  \item{Lat}{Latitude, numeric}
  \item{Lon}{Longitude, numeric}
  \item{Cls}{Climate Zone, factor with 30 levels}
}}
\source{
\url{http://koeppen-geiger.vu-wien.ac.at/present.htm}
}
\usage{
climatezones
}
\description{
A data frame containing the Koppen Geiger climate classification for areas.
}
\keyword{datasets}
