% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCoords.R
\name{genCoords}
\alias{genCoords}
\title{genCoords}
\usage{
genCoords(latlong = "lat", full = FALSE, latdim = 6480, londim = 12960)
}
\arguments{
\item{latlong}{Whether a vector of latitude values, or longitude values, is being created.}

\item{full}{Specify whether to generate a full list of all coords in the full image, or just one row/column (default FALSE, means just one row/column).}

\item{latdim}{The number of pixels in the image in the latitude dimension (y)}

\item{londim}{The number of pixels in the image in the longitude dimension (x)}
}
\value{
A vector of latitude or longitude coordinate values (default for a map 12960x 6480y).
}
\description{
This function will generate a list of coordinates for the 'fine' resolution climate zone map data, stored in 'kgz.rda'.
}
\examples{
ccoords <- genCoords(latlong='lon',full='true')
}
