% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZone.R
\name{getZone}
\alias{getZone}
\title{getZone}
\usage{
getZone(num)
}
\arguments{
\item{num}{The numeric code describing a climate zone.}
}
\value{
A string describing the climate zone as a character code.
}
\description{
This function will return the character code of a given climate zone corresponding to a numeric value (factor levels of the fine resolution data).
}
\examples{
ccz <- getZone(16)
}
