% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgcities.R
\docType{data}
\name{kgcities}
\alias{kgcities}
\title{Koppen Geiger climates for selected cities reference table.}
\format{A data frame with 88 rows and 5 variables:
\describe{
  \item{loc}{location, character}
  \item{rczd}{reported climate zone description, character}
  \item{rcz}{reported climate zone, character}
  \item{lon}{longitude, numeric}
  \item{lat}{latitude, numeric}
}}
\source{
\url{http://www.wikipedia.org}
}
\usage{
kgcities
}
\description{
A data frame containing the reported climate zones and longitude and latitude for selected example cities worldwide.
}
\keyword{datasets}
