% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmz.R
\docType{data}
\name{kmz}
\alias{kmz}
\title{High resolution (100s) Koppen Geiger climate zones.}
\format{A vector of length 83980800 (ydim=6480,xdim=12960, row-wise), with 32 factor levels representing 31 koppen-geiger climate zones and 'ocean'.}
\source{
Rubel, F., Brugger, K., Haslinger, K., Auer, I., 2016. The climate of the European Alps: Shift of very high resolution Köppen-Geiger climate zones 1800–2100. Meteorologische Zeitschrift. doi:10.1127/metz/2016/0816 \url{http://koeppen-geiger.vu-wien.ac.at/}
}
\usage{
kmz
}
\description{
A vector containing all points of a high resolution climate zone map.
}
\keyword{datasets}
