% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoundCoordinates.R
\name{RoundCoordinates}
\alias{RoundCoordinates}
\title{RoundCoordinates}
\usage{
RoundCoordinates(number, res = "course", latlong = "lat")
}
\arguments{
\item{number}{The value to round}

\item{res}{Specify the resolution as a string - 'course' for lower resolution (default, 0.5 degrees), 'fine' for higher resolution (100s).}

\item{latlong}{Whether the number is a latitude value ('lat', default), or longitude value ('lon') (this parameter is only used if res='fine').}
}
\value{
The inputed number, rounded to a value ending in either .25 or .75 (course), or to nearest 'fine' (100s) resolution grid point.
}
\description{
This function will round the inputed value to a co-ordinate present in the KG Climate Zone look up table.
}
\examples{
RoundCoordinates(10.98)
RoundCoordinates(10.98,res='fine',latlong='lon')
}
