% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunExample.R
\name{RunExample}
\alias{RunExample}
\title{RunExample}
\usage{
RunExample()
}
\value{
launches the shiny app created for the package.
}
\description{
This function will run the shiny app found in this package.
}
\examples{
\dontrun{
kgc::RunExample()
}
}
