% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranslateZipCode.R
\name{TranslateZipCode}
\alias{TranslateZipCode}
\title{TranslateZipCode}
\usage{
TranslateZipCode(number)
}
\arguments{
\item{number}{A dataframe that contains zip codes in a column labeled 'zip'}
}
\value{
The original data frame, with two additional columns for the corresponding latitude and longitude
}
\description{
This function will return the data frame with the longitude and latitude of the zip codes
}
\examples{
example <- data.frame("zip" = c(44106,638, 1106))
TranslateZipCode(example)
}
