% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipcodes.R
\docType{data}
\name{zipcodes}
\alias{zipcodes}
\title{Zip Code reference table.}
\format{A data frame with 33144 rows and 3 variables:
\describe{
  \item{zip}{zip code, numeric}
  \item{lat}{latitude, numic}
  \item{long}{long, numeric}
}}
\source{
\url{http://www.census.gov/geo/maps-data/data/gazetteer2016.html}
}
\usage{
zipcodes
}
\description{
A data frame containing the longitude and latitude for zip codes.
}
\keyword{datasets}
